fixed cancel delete button

+++ public/js/admin_profile.js
@@ -286,23 +286,26 @@ const deleteBtn = document.getElementById("deleteBtn");
 
 deleteBtn.addEventListener('click', function (e) {
     e.preventDefault();
-    confirm("Are you sure you want to delete your account?");
+    const res = confirm("Are you sure you want to delete your account?");
     console.log("clicked");
+    
+    if(res == 1){
+        let result = fetch("/deleteUser", {
+            method: "DELETE",
+            headers: {
+                "Content-Type": "application/json",
+            },
+            body: JSON.stringify({ "email": credEmail }),
+        }).then((response) => {
+            if (response.status == 200) {
+                window.location.href = "/html/index.html";
+            }
+            else {
+                alert("Error deleting account");
+            }
+        });
+    }
 
-    let result = fetch("/deleteUser", {
-        method: "DELETE",
-        headers: {
-            "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ "email": credEmail }),
-    }).then((response) => {
-        if (response.status == 200) {
-            window.location.href = "/html/index.html";
-        }
-        else {
-            alert("Error deleting account");
-        }
-    });
 
 });
